actualizacion sin recarga, middleware, contrasenas

+++ routes/web.php
@@ -16,19 +16,16 @@ Route::get('iniciar_sesion', [PagesController::class, 'login']) -> name('login')
 Route::post('iniciar_sesion', [LoginController::class, 'login'])-> name('login.post');
 Route::get('/logout', [LoginController::class, 'logout']) -> name('logout');
 
-Route::get('index_usuario', [PagesController::class, 'logged']) -> name('logged');
-Route::get('index_admin', [PagesController::class, 'logged_admin'])->name('logged_admin');
+Route::get('/lector', [AccesosController::class, 'index']) -> name('lector')->middleware('auth');
+Route::post('/lector', [AccesosController::class, 'store']) -> name('lector')->middleware('auth');
 
-Route::get('/lector', [AccesosController::class, 'index']) -> name('lector');
-Route::post('/lector', [AccesosController::class, 'store']) -> name('lector');
+Route::get('/espacios', [EspaciosController::class, 'index']) -> name('espacios')->middleware('auth');
+Route::post('/espacios', [EspaciosController::class, 'store']) -> name('espacios.store')->middleware('auth');
+Route::delete('/espacios/{espacio}', [EspaciosController::class, 'destroy']) -> name('espacios.destroy')->middleware('auth');
+Route::put('/espacios/{espacio}', [EspaciosController::class, 'update']) -> name('espacios.update')->middleware('auth');
 
-Route::get('/espacios', [EspaciosController::class, 'index']) -> name('espacios');
-Route::post('/espacios', [EspaciosController::class, 'store']) -> name('espacios.store');
-Route::delete('/espacios/{espacio}', [EspaciosController::class, 'destroy']) -> name('espacios.destroy');
-Route::put('/espacios/{espacio}', [EspaciosController::class, 'update']) -> name('espacios.update');
-
-Route::get('ver-estudiantes', [EstudiantesController::class, 'index']) -> name('ver-estudiantes');
-Route::post('ver-estudiantes', [EstudiantesController::class, 'filter']) -> name('ver-estudiantes.filter');
+Route::get('ver_estudiantes', [EstudiantesController::class, 'index']) -> name('ver-estudiantes')->middleware('auth');
+Route::post('ver_estudiantes', [EstudiantesController::class, 'filter']) -> name('ver-estudiantes.filter')->middleware('auth');
 
 Route::get('politicas', [PagesController::class, 'politicas']) -> name('politicas');
